Update to how the form behaves.
Removed the timer which brings the form to the top most position on screen. Found out that there is an api call that handles the top most function also found some other extended styles for making the form not appear in the alt + tab selection window. There was an issue when I would run the program it would become the active program, there was an option within the styles to remedy that. So added it in.